Implement stat() nanosecond resolution where it's supported.
This commit implements nanosecond resolution via three new fields in stat()'s
return value, and also changes fstat() and lstat(). Support is also added to
streamWrapper::stream_stat() and streamWrapper::url_stat() for the new fields,
and this is used by the php://memory and php://temp wrappers.

This is presently untested on Windows, and I suspect there's a better approach
possible, which is why I'm holding off on committing this to master right now.